requirements.txt
simple-salesforce
stomp.py
pyyaml
flask